Add header to index.html with styles in style.css
Used flexbox to get the layout looking right. Always feels like I'm
overdoing it with divs, but the header needed to be flex, and needed a
box inside it to contain the logo and the links. I think I could have
made it a little simpler by including the logo in the ul, rather than
as it's own element. Maybe a change in the future?