Style body and header
I finally figured out what I was doing wrong, and was able to simplify
the header by removing the nav div and moving the header logo text into
the header ul. I used align-self to get the links and logo aligned, and
just set the ul itself to width: 900px.