Style the hero section
I always feel a little funny having so many nested divs - I think it
can be difficult to read sometimes - but I think it was nececssary
here. The hero section is an item of the body container, and is also a
container itself. Inside the hero container are two items: text (and a
buttton) on the left and an image on the right. The text and button
need to be in a div, as they need to function as one item. The image
can stand alone, as it is just one item. The hero-text div doesn't
need to flex, as everything inside it just left-aligns.

I struggled a bit with the size and spacing, but solved this by setting
the max-width of the hero-text div to 400px (I have the main width of
each section set at 900px) and then using "justify-content:
space-between" on the hero-main container.