Style the call to action section
Styled the div within the main action section div as a blue box with a
div on the left (flex column) and a button on the right. Used padding
to get the spacing, centered the div and button on the cross (vertical)
axis, and spaced evenly on the main axis.