Add padding to top and bottom of sections
I didn't bother with this while focusing on the main styles, and held
back to do it all at the end. I'd thought they could all be the same,
but I don't think the spacing is the same top and bottom. I did group
most of the sections together to set the widths to 100%, all but the
call to action section, as it doesn't need it.